backend/requirements.txt
flask==2.0.1
flask-cors==3.1.1
pytesseract==0.3.8
pdfminer.six==20201018
gunicorn==20.1.0  # For production deployment (Heroku)
